+++ SimpleCDemos/DemoLua548/elua-0.9/src/lua/lmem.h
@@ -9,6 +9,7 @@
 
 
 #include <stddef.h>
+#include <stdint.h>  /* For uint32_t */
 
 #include "llimits.h"
 #include "lua.h"
@@ -35,12 +36,12 @@
    ((v)=cast(t *, luaM_reallocv(L, v, oldn, n, sizeof(t))))
 
 
-LUAI_FUNC void *luaM_realloc_ (lua_State *L, void *block, size_t oldsize,
-                                                          size_t size);
-LUAI_FUNC void *luaM_reallocv_impl (lua_State *L, void *block, int oldn, int n, size_t size_elem);
+LUAI_FUNC void *luaM_realloc_ (lua_State *L, void *block, uint32_t oldsize,
+                                                          uint32_t size);
+LUAI_FUNC void *luaM_reallocv_impl (lua_State *L, void *block, int oldn, int n, uint32_t size_elem);
 LUAI_FUNC void *luaM_toobig (lua_State *L);
 LUAI_FUNC void *luaM_growaux_ (lua_State *L, void *block, int *size,
-                               size_t size_elem, int limit,
+                               uint32_t size_elem, int limit,
                                const char *errormsg);
 
 #endif
