+++ SimpleCDemos/DemoLua548/elua-0.9/src/lua/lzio.h
@@ -9,6 +9,7 @@
 #define lzio_h
 
 #include "lua.h"
+#include <stdint.h>  /* For uint32_t */
 
 #include "lmem.h"
 
@@ -47,7 +48,7 @@ typedef struct Mbuffer {
 #define luaZ_direct_mode(zio) (luaZ_get_base_address(zio) != NULL)
 #define luaZ_get_crt_address(zio) (luaZ_get_base_address(zio) + (zio)->i)
 
-LUAI_FUNC char *luaZ_openspace (lua_State *L, Mbuffer *buff, size_t n);
+LUAI_FUNC char *luaZ_openspace (lua_State *L, Mbuffer *buff, uint32_t n);
 LUAI_FUNC void luaZ_init (lua_State *L, ZIO *z, lua_Reader reader,
                                         void *data);
 LUAI_FUNC size_t luaZ_read (ZIO* z, void* b, size_t n);	/* read next n bytes */
